entity manager & messagedispatcher

+++ artificial_intelligence/assignment2/Telegram.h
@@ -6,10 +6,10 @@ struct Telegram
 	int sender;
 	int receiver;
 	MessageType *msg;
-	double dispatchTime;
+	long long dispatchTime;
 	void* extraInfo;
 
-	Telegram(const int &sender, const int &receiver, MessageType &msg, const double &distpatchTime, void *extraInfo = nullptr) : sender(sender), receiver(receiver), msg(&msg), dispatchTime(distpatchTime), extraInfo(extraInfo) {}
+	Telegram(const int &sender, const int &receiver, MessageType &msg, const long long &distpatchTime, void *extraInfo = nullptr) : sender(sender), receiver(receiver), msg(&msg), dispatchTime(distpatchTime), extraInfo(extraInfo) {}
 	bool operator<(const Telegram &other) const
 	{
 		return dispatchTime < other.dispatchTime;
